==> draft_notebooks/0_Data_sources.ipynb <==
# Extracting and processing data streams

The overall idea of this notebook is to use it to start extracting the different data streams that the [Interactive Atlas of Heart Disease and Stroke](https://nccd.cdc.gov/DHDSPAtlas) used to make its different choropleth maps.

### Target variable

The CDC makes the multiple mortality cause file for each year available for public download [here](https://www.cdc.gov/nchs/data_access/vitalstatsonline.htm#Mortality_Multiple). It looks like this [https://gist.github.com/SohierDane/0f2cf7a8538ca35431dd7575ac38e7ca](python script) here is able to extract the data from these fixed-width data files, although I'm not entirely sure how exports the data; as json? CSVs?

Target variable source: CDC multiple cause of mortality file
To do:
* Need to download the various fixed-width data files to AWS (let's set 2000 to 2016 as the years that you're interested in)
* Need to figure out how to extract these files into CSVs and then to load them into postgres or some other database structure, so that you can query it.

import csv
import pandas as pd

### Predictor variables

To start off, let's see if we can pick 3 'good' predictor variables. What I mean by 'good' is that I can get time series data for each county in the US (or at least a significant subset of them), at a reasonable sampling frequency. E.g., to at least be able to have a time point annually, if not bi-annually or even monthly.

Let's just pick some time streams that I think would be interesting; I'm not sure that this information is going to be available at the county level, though:
* Percent diagnosed with diabetes
* Percent with health insurance
* Unemployment rate (vs poverty rate vs income inequality)
* Blood pressure medication non-adherence

What you might have to do is to reformulate this to do a state-by-state basis, because there's probably more data available on the state level (and possibly at a higher frequency) than at the local level. The other upside is that when public health interventions are implemented, they are often implemented at the state level, and finding news articles about these interventions is probably easier for state-wide actions vs county-level actions. The downside, though, is that states are clearly extremely different sizes, and so the model might get confused or act differently between predicting mortality or disease incidence in Rhode Island vs California.

#### Diabetes predictor variables

From the CDC: diabetes prevalence, incidence, obesity, and leisure-time physical inactivity by county. Data is from 2004 to 2013. Datasets located at [this website](https://www.cdc.gov/diabetes/data/countydata/countydataindicators.html). After reading further, it looks like this data actually comes from the BRFSS? Since the raw BRFSS data is actually available, it might be better/necessary to directly process the BRFSS dataset.

#### BRFSS

The Behavioral Risk Factor Surveillance System is a data set that I think has a lot of variables that would potentially be interesting, and I think would give me the percent diagnosed with diabetes and also the percent diagnosed with health insurance, but it looks like cleaning up the data is fairly involved. It does seem that this guy has already done a lot of the heavy lifting at his github [here](https://github.com/winstonlarson/brfss), but I'm going to have to poke around the data in order to see if it actually provides county-level data.

%%bash
ls ../data/brfss

The brfss2014 csv is quite large, so we'll just get the first line, which is the header for the columns.

with open('../data/brfss/brfss2014.csv', newline='') as f:
    reader = csv.reader(f)
    row1 = next(reader)
    row2 = next(reader)

brfss_2014_sample = pd.read_csv("../data/brfss/brfss2014.csv", nrows=100000)

list(brfss_2014_sample.columns)

brfss_2014_sample['x.state'].unique()

groupby_state = brfss_2014_sample.groupby(['x.state'])[['fmonth', 'idate', 'imonth', 'iday', 'iyear', 'hlthpln1', 'medcost']]

groupby_state.get_group(1)

brfss_2014 = pd.read_csv("../data/brfss/brfss2014.csv", encoding = "ISO-8859-1", engine='python', nrows= 1000)

brfss_2014

%%bash
cd ../data/brfss
wc -l brfss2014.csv
