==> draft_notebooks/3_BRFSS_data_(predictors).ipynb <==
## Get variable names and descriptions from BRFSS codebooks

import pandas as pd
import pickle

# Get pandas and postgres to work together
import psycopg2 as pg
import pandas as pd
import pandas.io.sql as pd_sql

import re
import PyPDF2
import pickle

from progress_bar import log_progress
from functools import reduce

First, let's load one of the CSVs and take a look at the data structure.

brfss_2014 = pd.read_csv("../data/brfss/brfss2014.csv", encoding = "cp1252", nrows=100)

brfss_2014.head()

## Extracting relevant predictors from BRFSS 1999-2016

BRFSS has a lot different variables, and unfortunately the name of the variables changes over the years. We need to extract the names of the variables and the description of each variable from the codebooks for each year. This is going to involve reading in each of the codebook pdf files and extracting the variable names and the description of the question that was asked for each variable. ONce we have a dictionary for the variable names and the questions, we can figure out which variables we want to get from each BRFSS CSV file.

def extract_pdf_string(pdf_path):
    
    pdfFileObj = open(pdf_path, 'rb')
    
    pdfReader = PyPDF2.PdfFileReader(pdfFileObj) 
    num_pages =  pdfReader.numPages
    
    pdf_string = ""
    
    for n in range(num_pages):
        pageObj = pdfReader.getPage(n)
        page_string = pageObj.extractText()
        clean_page_string = re.sub("\s+", " ", page_string)
        clean_page_string = clean_page_string.strip()
        pdf_string += clean_page_string
        
    # closing the pdf file object
    pdfFileObj.close()
    return pdf_string

def find_var(words):
    for word in words:
        if word == 'BRFSS' or word == 'SAS' or len(word) <3:
            pass
        elif word[0] == '_':
            return word
        else:
            ratio = sum([letter.isupper() for letter in list(word)])/len(list(word))
            if ratio > 0.6:
                return word

def extract_variable_name_and_description(section):
    words = section.split(" ")
    var_name = find_var(words)
    
    if 'Description:' in section:
        subsections = section.split("Description:")
    elif 'Question:' in section:
        subsections = section.split("Question:")
    else:
        description = None
        return var_name, description
        
    description = subsections[1]

    value_limit = description.find("Value")
    weighted_limit = description.find("Weighted")

    if value_limit == -1:
        limit = weighted_limit
    elif weighted_limit == -1:
        limit = value_limit
    else:
        limit = min(description.find("Value"), description.find("Weighted"))

    description = description[0:limit]
    description = description.strip()

    if description == '':
        description = None
    
    return var_name, description

years = range(1999, 2018)

codebook_dfs_dict = {}

for year in log_progress(years):
    pdf_string = extract_pdf_string(f"../data/brfss/codebooks/{year}_codebook.pdf")
    pdf_sections = pdf_string.split("SAS Variable")

    var_desc_array = []
    for section in pdf_sections:
        row = extract_variable_name_and_description(section)
        var_desc_array.append(row)

    var_desc_df = pd.DataFrame(var_desc_array, columns=['var_name', 'description'])
    
    # We'll drop the rows where the variable is none, and then also get rid of duplicate rows where the var_names
    # are the same.
    var_desc_df = var_desc_df.dropna(subset=['var_name']).drop_duplicates(subset=['var_name'])

    codebook_dfs_dict[year] = var_desc_df

with open("../data/pickles/codebook_dfs_dict.pkl", "wb") as f:
    pickle.dump(codebook_dfs_dict, f)

We now have a dictionary of codebook dataframes for each year; the codebook dataframe has a column 'var_name' and a column 'description', and each row is a different variable. Now, let's try joining these dataframes together and seeing how it goes. The codebooks have different variable names for the same information between different years, which unfortunately means that in order to see how these names have evolved we have to use an outer join to get master dataframe of all of the variable names and their descriptions over the years.

master_df = codebook_dfs_dict[1999]
master_df.rename(columns={'description':1999}, inplace=True)
for year, df in codebook_dfs_dict.items():
    if year == 1999:
        pass
    else:
        master_df = pd.merge(master_df, df, how='outer', on='var_name')
        master_df.rename(columns={'description':year}, inplace=True)

with open("../data/pickles/master_codebook_all_years.pkl", "wb") as f:
    pickle.dump(master_df, f)

master_df.shape

master_df[master_df.var_name=='_LLCPWT']

master_df[master_df.var_name=='_FINALWT']

We could do a merge with an inner join, which would not capture every variable name that was ever used over all the codebooks (which an outer join does).

intersection_df = codebook_dfs_dict[1999]
intersection_df.rename(columns={'description':1999}, inplace=True)
for year, df in codebook_dfs_dict.items():
    if year == 1999:
        pass
    else:
        intersection_df = pd.merge(master_df, df, on='var_name')
        intersection_df.rename(columns={'description':year}, inplace=True)

intersection_df.head()

We can keep only the variables where 5 or fewer codebooks did not have that variable; that is, most of the codebooks contain that variable name.

majority_vars = intersection_df[intersection_df.isna().sum(axis=1) <= 5]

We want to translate the master dataframe of all the variable names and their descriptions (with each year as a different column) into a dictionary. What we're interested in is the variable name as a key, and the most 'common' description as the value. The description/question wording changes between years, so we'll pick the description/question that has been used most repeatedly over the years.

def make_consensus_var_desc_dict(df):
    raw_dict = dict(pd.DataFrame.transpose(df.set_index('var_name')))
    clean_dict = {}
    for key, value in raw_dict.items():
        if value.mode().empty:
            clean_dict[key] = None
        else:
            clean_dict[key] = value.mode()[0]
    return clean_dict

It actually makes the most sense to make a consensus variable/description dictionary from the dataframe generated by merging with outer join, because then we can look through the key/value pairs and figure out which variables are synonyms with each other.

consensus_var_desc_dict = make_consensus_var_desc_dict(master_df)

with open("../data/pickles/consensus_var_desc_dict.pkl", "wb") as f:
    pickle.dump(consensus_var_desc_dict, f)

for key, value in d.items():
    if 'DIABE' in key:
        print(key)
        print(value)

master_df[master_df.var_name.str.contains("DIABET") == True]

## Reading brfss CSVs and extracting relevant variables.

brfss_1999_df = pd.read_csv("../data/brfss/csv/brfss1999.csv")

brfss_1999_df.head()

diabetes_columns = [col for col in brfss_1999_df.columns if 'diabe' in col]

diabetes_columns

for col in brfss_1999_df.columns:
    if 'diabe' in col:
        z = col

columns_to_extract = ['FINALWT', 'DIABET', 'LLCPWT', 'MEDCOST', 'CHOLC']

columns_to_extract = ['SMOKE', 'FINALWT', 'LLCPWT']

columns_to_extract = ['CHOLC', 'TOLDHI2', 'FINALWT', 'LLCPWT']

columns_to_extract = ['LASTCHNG', 'ASTHMA2', 'ASTHMA3', 'FINALWT', 'LLCPWT']

columns_to_extract = ['STATE', 'EXERANY', 'FINALWT', 'LLCPWT']

for col in columns_to_extract:
    print(col.lower())

def find_cols_this_df_has(list_of_columns, df):
    good_cols = []
    for col in df.columns:
        temp_col = col.replace("x.", "")
        for col_to_extract in list_of_columns:
            if col_to_extract.lower() in temp_col.lower():
                good_cols.append(col)
    return good_cols

find_cols_this_df_has(columns_to_extract, brfss_1999_df)

d['CHOLC']

# Okay, let's go basic. Let's decide what we want to extract from each csv. How about let's extract the variable
# smoke100 and the weight from every csv.
list_of_temp_dfs = []

for year in log_progress(years):
    brfss_df = pd.read_csv(f"../data/brfss/csv/brfss{year}.csv", encoding="cp1252")#, nrows=10)
    cols = find_cols_this_df_has(columns_to_extract, brfss_df)
#     print(year)
#     print(cols)
    temp_df = brfss_df[cols].copy()
    
    temp_df['year'] = year
    list_of_temp_dfs.append(temp_df)

with open("../data/pickles/any_exercise_list_of_dfs.pkl", "wb") as f:
    pickle.dump(list_of_temp_dfs, f)

print(list_of_temp_dfs[0].exerany.isna().sum())
print(len(list_of_temp_dfs[0].exerany))
list_of_temp_dfs[0].exerany.value_counts()
