==> draft_notebooks/8_scraping_targets.ipynb <==
from selenium import webdriver
from selenium.webdriver.common.keys import Keys
from selenium.webdriver.support.ui import Select
import time

import os
chromedriver = "/Applications/chromedriver" # path to the chromedriver executable
os.environ["webdriver.chrome.driver"] = chromedriver

driver = webdriver.Chrome(chromedriver)

cdc_wonder_mcd = "https://wonder.cdc.gov/mcd-icd10.html"
driver.get(cdc_wonder_mcd)

agree_button = driver.find_element_by_name("action-I Agree")
agree_button.click()
time.sleep(2)

# Select group by option for results as state
select_group_results = Select(driver.find_element_by_id("SB_1"))
select_group_results.select_by_visible_text("UCD - ICD Sub-Chapter")

select_group2_results = Select(driver.find_element_by_id("SB_2"))
select_group2_results.select_by_visible_text("Month")

def search_dict_values(dic, search_for):
    for key, value in dic.items():
        if search_for in value:
            return key

select_location = Select(driver.find_element_by_id("codes-D77.V9"))

location_dict = {}
for option in select_location.options:
    location_dict[option.get_attribute('value')] = option.text

# let's get the location values for the first 5 states.
location_values = [search_dict_values(location_dict, state) for state in states_by_five[0]]
#location_value = search_dict_values(location_dict, "Alabama")
for location_value in location_values:
    select_location.select_by_value(location_value)

select_location.deselect_by_value("*All*")

# Click checkbox to show zero rows.
show_zeros_button = driver.find_element_by_id("CO_show_zeros")
show_zeros_button.click()

# Click checkbox to show suppressed rows.
show_suppressed_button = driver.find_element_by_id("CO_show_suppressed")
show_suppressed_button.click()

# Submit this request to the last send button on the page, so that all of the criteria are included.
send_button = driver.find_element_by_xpath("//div[@class='footer-buttons']/input")
#print(send_button.get_attribute('title'))
send_button.click()

# It takes time for CDC wonder to answer the query and display the results page.
time.sleep(20)
